timer: remove bogus feature on writing to count register
NOTE: Actually do this commit upon seeing writing to the count registers
of timers are invalid. The count register timer-example.v is WRITABLE, but the
document on the function of it DOES NOT SAY SO. On the contrary, this situation
is DECLARED EXCEPTIONAL in cscore.net, and therefore is implied as such.