src: synthesis warning elimination
F**K. The f**king noisy HDL synthesizer does not understand anything
relevant about good software design properties and starts to optimize
this and that and starts to spill out a thousand lines of warnings and
nearly caught fire and smoked. So I have to find out what do the two
messages that is seemingly random describing a missed signal of clk and
rst. And I found almost nothing wrong, just two wrongly named pins in
the top-level module. Strange. Disappointing. Bullying. Why don't you
behave like a almost modern compiler and actually *silently* optimize
anything? It's not cool when in 2019 we have to write Verilog in the
fashion of assembler (tm) when it has to be synthesized on your devices.
Why on earth don't you be a good tame smart optimizing compiler just
like the software land did? Why the f**k do you need to let *me*
optimize anything and everything and reduce the human-friendlyness (not
to mention user-friendlyness) to gibberish? Stupid shit xst.